Update patch set 1
Patch Set 1:

Please help me understand why this is all in the docs-templates directory? 
It will never be rendered on read the docs unless it is in the docs/ directory
so how are users going to find it?

Patch-set: 1
Reviewer: Gerrit User 1000089 <1000089@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Verified=0