Update patch set 1
Patch Set 1: Code-Review-1

Changing my vote to ask a very specific question.

Aric you almost certainly know the answer.

The verify job result is left in an archive reachable from the logs directory. Here's one of them:

https://logs.o-ran-sc.org/production/vex-sjc-oran-jenkins-prod-1/rtdv3-global-verify-master/76/html/

Go visit that.  Check the link to portal-ric-dashboard. The mapping that is coded there uses the global URL, so the link doesn't go to a page within that generated area, but out to RTD.io.  I think the link should be local so you can test locally and validate that everything works BEFORE merge.

How do we write that kind of link, so it works locally AND when pushed to RTFD.io?

Please advise.

Patch-set: 1