Repairs for doc repo.
On this change:
default-version is a variable that is set in
ci-management/jjb/rtdv3/rtdv3.yaml
it should only be set once all projects have branched.

version: on the master branch this should stay master
further, now that we have changed read the docs themes
we don't need to touch this any more.

On deleting and re-creating the bronze branch.

something has happened with the bronze branch
I think it was created twice at some point
perhaps once like "Bronze" and once as "bronze"

This has confused read the docs, which is now referring
to the bronze branch as bronze_a which we do not want.

I would like to delete the bronze branch and re-create it at the ref
created by this change.
this will allow me to fix read the docs, specifically the version
switcher.

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I4067c09412d77efaad7f5fc67359ab7c01628a5b